requirements.txt
torch
torchvision
numpy
Pillow
open_clip_torch
safetensors
einops
wandb
